fix(critical): restore Goal.Id to long type for production compatibility
- Goal.Id was changed from long to int in v2.8.0, causing OverflowException in production
- Production database contains old goal IDs as large timestamps (milliseconds since epoch)
- Restore Goal.Id, IGoalService, GoalService, FinanceService, Goals.razor to use long
- Update tests to use long literal (1L) for FindAsync calls
- All 41 tests passing

Closes: OverflowException at Microsoft.Data.Sqlite.SqliteDataReader.GetInt32

+++ FamilyFinance/Services/FinanceService.cs
@@ -85,7 +85,7 @@ public class FinanceService
     public Task SaveGoalAsync(Goal goal) 
         => _goals.SaveAsync(goal);
 
-    public Task DeleteGoalAsync(int id) 
+    public Task DeleteGoalAsync(long id) 
         => _goals.DeleteAsync(id);
 
     // Portfolios - delegate to IPortfolioService
